Moved Update/Delete Event logic to EventDetail.
EventCard is now just a presentational component.
update and remove event functions moved from EventList to EventDetail.
